Separate Shape Handling Code and Integrate Functionality
